requirements.txt
pdfminer.six
genanki
deep_translator
